refactor(action): move customViewAction to new Revision/ActionController
