+++ EMS/core-bundle/src/Controller/ContentManagement/ActionController.php
@@ -12,8 +12,8 @@ use EMS\CoreBundle\Entity\Template;
 use EMS\CoreBundle\Form\Data\BoolTableColumn;
 use EMS\CoreBundle\Form\Data\EntityTable;
 use EMS\CoreBundle\Form\Data\TableAbstract;
-use EMS\CoreBundle\Form\Form\TableType;
 use EMS\CoreBundle\Form\Form\ActionType;
+use EMS\CoreBundle\Form\Form\TableType;
 use EMS\CoreBundle\Helper\DataTableRequest;
 use EMS\CoreBundle\Repository\ContentTypeRepository;
 use EMS\CoreBundle\Repository\TemplateRepository;
