+++ EMS/core-bundle/src/Controller/ContentManagement/DataController.php
@@ -5,9 +5,6 @@ namespace EMS\CoreBundle\Controller\ContentManagement;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\NoResultException;
 use EMS\CommonBundle\Helper\EmsFields;
-use EMS\CommonBundle\Service\Pdf\Pdf;
-use EMS\CommonBundle\Service\Pdf\PdfPrinterInterface;
-use EMS\CommonBundle\Service\Pdf\PdfPrintOptions;
 use EMS\CoreBundle\Core\ContentType\ContentTypeRoles;
 use EMS\CoreBundle\Core\ContentType\ViewTypes;
 use EMS\CoreBundle\Core\Log\LogRevisionContext;
@@ -24,13 +21,10 @@ use EMS\CoreBundle\Entity\View;
 use EMS\CoreBundle\Exception\DuplicateOuuidException;
 use EMS\CoreBundle\Exception\ElasticmsException;
 use EMS\CoreBundle\Form\Field\IconTextType;
-use EMS\CoreBundle\Form\Field\RenderOptionType;
 use EMS\CoreBundle\Form\Form\RevisionType;
 use EMS\CoreBundle\Helper\EmsCoreResponse;
 use EMS\CoreBundle\Repository\ContentTypeRepository;
-use EMS\CoreBundle\Repository\EnvironmentRepository;
 use EMS\CoreBundle\Repository\RevisionRepository;
-use EMS\CoreBundle\Repository\TemplateRepository;
 use EMS\CoreBundle\Routes;
 use EMS\CoreBundle\Service\ContentTypeService;
 use EMS\CoreBundle\Service\DataService;
@@ -46,7 +40,6 @@ use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -54,12 +47,21 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Validator\Constraints\Regex;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Twig\Environment as TwigEnvironment;
-use Twig\Error\Error;
 
 class DataController extends AbstractController
 {
-    public function __construct(private readonly LoggerInterface $logger, private readonly DataService $dataService, private readonly SearchService $searchService, private readonly ContentTypeService $contentTypeService, private readonly EnvironmentService $environmentService, private readonly IndexService $indexService, private readonly TranslatorInterface $translator, private readonly ViewTypes $viewTypes, private readonly TwigEnvironment $twig, private readonly PdfPrinterInterface $pdfPrinter, private readonly JobService $jobService)
-    {
+    public function __construct(
+        private readonly LoggerInterface $logger,
+        private readonly DataService $dataService,
+        private readonly SearchService $searchService,
+        private readonly ContentTypeService $contentTypeService,
+        private readonly EnvironmentService $environmentService,
+        private readonly IndexService $indexService,
+        private readonly TranslatorInterface $translator,
+        private readonly ViewTypes $viewTypes,
+        private readonly TwigEnvironment $twig,
+        private readonly JobService $jobService
+    ) {
     }
 
     public function rootAction(string $name): Response
@@ -513,126 +515,6 @@ class DataController extends AbstractController
         return $viewType->generateResponse($view, $request);
     }
 
-    public function customViewAction(string $environmentName, int $templateId, string $ouuid, bool $_download, bool $public): Response
-    {
-        /** @var EntityManager $em */
-        $em = $this->getDoctrine()->getManager();
-
-        /** @var TemplateRepository $templateRepository */
-        $templateRepository = $em->getRepository(Template::class);
-
-        /** @var Template|null $template * */
-        $template = $templateRepository->find($templateId);
-
-        if (null === $template || ($public && !$template->isPublic())) {
-            throw new NotFoundHttpException('Template type not found');
-        }
-
-        /** @var EnvironmentRepository $environmentRepository */
-        $environmentRepository = $em->getRepository(Environment::class);
-
-        $environment = $environmentRepository->findBy([
-            'name' => $environmentName,
-        ]);
-
-        if (!$environment || 1 != \count($environment)) {
-            throw new NotFoundHttpException('Environment type not found');
-        }
-
-        /** @var Environment $environment */
-        $environment = $environment[0];
-
-        $document = $this->searchService->get($environment, $template->giveContentType(), $ouuid);
-
-        try {
-            $body = $this->twig->createTemplate($template->getBody());
-        } catch (Error $e) {
-            $this->logger->error('log.template.twig.error', [
-                'template_id' => $template->getId(),
-                'template_name' => $template->getName(),
-                'template_label' => $template->getLabel(),
-                'error_message' => $e->getMessage(),
-            ]);
-            $body = $this->twig->createTemplate($this->translator->trans('log.template.twig.error', [
-                '%template_id%' => $template->getId(),
-                '%template_name%' => $template->getName(),
-                '%template_label%' => $template->getLabel(),
-                '%error_message%' => $e->getMessage(),
-            ], EMSCoreBundle::TRANS_DOMAIN));
-        }
-
-        if (RenderOptionType::PDF === $template->getRenderOption() && ($_download || !$template->getPreview())) {
-            $output = $body->render([
-                'environment' => $environment,
-                'contentType' => $template->getContentType(),
-                'object' => $document,
-                'source' => $document->getSource(),
-                '_download' => true,
-            ]);
-            $filename = $this->generateFilename($this->twig, $template->getFilename() ?? 'document.pdf', [
-                'environment' => $environment,
-                'contentType' => $template->getContentType(),
-                'object' => $document,
-                'source' => $document->getSource(),
-                '_download' => true,
-            ]);
-
-            $pdf = new Pdf($filename, $output);
-            $printOptions = new PdfPrintOptions([
-                PdfPrintOptions::ATTACHMENT => PdfPrintOptions::ATTACHMENT === $template->getDisposition(),
-                PdfPrintOptions::COMPRESS => true,
-                PdfPrintOptions::HTML5_PARSING => true,
-                PdfPrintOptions::ORIENTATION => $template->getOrientation() ?? 'portrait',
-                PdfPrintOptions::SIZE => $template->getSize() ?? 'A4',
-            ]);
-
-            return $this->pdfPrinter->getStreamedResponse($pdf, $printOptions);
-        }
-        if ($_download || (0 === \strcmp($template->getRenderOption(), RenderOptionType::EXPORT) && !$template->getPreview())) {
-            if (null != $template->getMimeType()) {
-                \header('Content-Type: '.$template->getMimeType());
-            }
-
-            $filename = $this->generateFilename($this->twig, $template->getFilename() ?? $ouuid, [
-                'environment' => $environment,
-                'contentType' => $template->getContentType(),
-                'object' => $document,
-                'source' => $document->getSource(),
-            ]);
-
-            if (!empty($template->getDisposition())) {
-                $attachment = ResponseHeaderBag::DISPOSITION_ATTACHMENT;
-                if ('inline' == $template->getDisposition()) {
-                    $attachment = ResponseHeaderBag::DISPOSITION_INLINE;
-                }
-                \header("Content-Disposition: $attachment; filename=".$filename.($template->getExtension() ? '.'.$template->getExtension() : ''));
-            }
-            if (null != $template->getAllowOrigin()) {
-                \header('Access-Control-Allow-Origin: '.$template->getAllowOrigin());
-                \header('Access-Control-Allow-Headers: Content-Type, Authorization, Accept, Accept-Language, If-None-Match, If-Modified-Since');
-                \header('Access-Control-Allow-Methods: GET, HEAD, OPTIONS');
-            }
-
-            $output = $body->render([
-                'environment' => $environment,
-                'contentType' => $template->getContentType(),
-                'object' => $document,
-                'source' => $document->getSource(),
-            ]);
-            echo $output;
-
-            exit;
-        }
-
-        return $this->render('@EMSCore/data/custom-view.html.twig', [
-            'template' => $template,
-            'object' => $document,
-            'environment' => $environment,
-            'contentType' => $template->getContentType(),
-            'body' => $body,
-        ]);
-    }
-
     public function customViewJobAction(string $environmentName, int $templateId, string $ouuid, Request $request): Response
     {
         $em = $this->getDoctrine()->getManager();
@@ -1049,21 +931,5 @@ class DataController extends AbstractController
                 $this->reorderCollection($elem);
             }
         }
-    }
-
-    /**
-     * @param array<string, mixed> $options
-     */
-    private function generateFilename(TwigEnvironment $twig, string $rawTemplate, array $options): string
-    {
-        try {
-            $template = $twig->createTemplate($rawTemplate);
-            $filename = $template->render($options);
-            $filename = \preg_replace('~[\r\n]+~', '', $filename);
-        } catch (\Throwable) {
-            $filename = null;
-        }
-
-        return $filename ?? 'error-in-filename-template';
     }
 }
